+++ src/Service/Lambda/src/Result/ListFunctionsResponse.php
@@ -381,6 +381,7 @@ class ListFunctionsResponse extends Result implements \IteratorAggregate
             'SubnetIds' => !isset($json['SubnetIds']) ? null : $this->populateResultSubnetIds($json['SubnetIds']),
             'SecurityGroupIds' => !isset($json['SecurityGroupIds']) ? null : $this->populateResultSecurityGroupIds($json['SecurityGroupIds']),
             'VpcId' => isset($json['VpcId']) ? (string) $json['VpcId'] : null,
+            'Ipv6AllowedForDualStack' => isset($json['Ipv6AllowedForDualStack']) ? filter_var($json['Ipv6AllowedForDualStack'], \FILTER_VALIDATE_BOOLEAN) : null,
         ]);
     }
 }