+++ src/Service/Rekognition/src/Enum/CustomizationFeature.php
@@ -1,0 +1,17 @@
+<?php
+
+namespace AsyncAws\Rekognition\Enum;
+
+final class CustomizationFeature
+{
+    public const CONTENT_MODERATION = 'CONTENT_MODERATION';
+    public const CUSTOM_LABELS = 'CUSTOM_LABELS';
+
+    public static function exists(string $value): bool
+    {
+        return isset([
+            self::CONTENT_MODERATION => true,
+            self::CUSTOM_LABELS => true,
+        ][$value]);
+    }
+}